chore: move normalize order and normalize transfer to orders and accounts libs
